Add events and results (#27)
Porting some events and results features over from NW5k. At the moment this is behaviour is not exposed through the navigation, you'll have to hit the routes directly (once I've added the results to the database I'll switch from the static results page to the dynamic ones).

Changes in this PR include:

 - Add Events controller and views for index and show (plus an action to show the latest event)
 - Add names and display info to Users (including tests)
 - Remove the static_pages layout and shift everything to the main layout
 - Improve the styling of the sign in page
 - Add back the gradient block on the home page
 - Tidy up the routes